dns-response-time-checker/requirements.txt
dnspython==2.3.0
numpy==1.23.4
pandas==1.5.0
python-dateutil==2.8.2
six==1.16.0
pytz==2022.7.1